runlock: rate limit cron job execution
runlock is a standalone command line program for supervising a task run
from a cron job. It can be used with any cron scheduler. For example,
using pseudocron (https://github.com/msantos/pseudocron):

    # run cron job hourly
    # do not run nc if the job succeeded in the last month
    pseudocron "3 * * * *" \
        runlock --timestamp="$(date --date="last month" +%s)" \
        nc -z example.com 80

Cron tasks are run in a new process group. If the elaspsed time of the
task exceeds the allotted runtime, the process group will be terminated
with a signal.

For a simpler way of scheduling tasks, runcron
(https://github.com/msantos/runcron) combines the cron scheduler with
task supervision.